Rename folder 'app' to 'music_lib_bot'
